修复获取礼物列表crash问题
修复获取礼物列表crash问题

+++ LinekeLive/LinekeLive/Classes/ModuleControllers/Player/Model/LKGiftModel.h
@@ -12,6 +12,6 @@
 
 @property (nonatomic, copy) NSString *img2; //图片
 @property (nonatomic, copy) NSString *subject; //名称
-@property (nonatomic, assign) double coin; //价格
+@property (nonatomic, assign) long coin; //价格
 
 @end
